+++ src/main/java/brawljars/api/intern/brawlers/BrawlerApiImpl.java
@@ -16,8 +16,8 @@
  */
 package brawljars.api.intern.brawlers;
 
-import brawljars.api.ApiContext;
-import brawljars.api.BaseApi;
+import supercell.api.wrapper.essentials.api.ApiContext;
+import supercell.api.wrapper.essentials.api.BaseApi;
 
 import java.util.concurrent.Future;