+++ src/main/java/brawljars/api/intern/clubs/info/ClubResponse.java
@@ -16,13 +16,13 @@
  */
 package brawljars.api.intern.clubs.info;
 
-import brawljars.common.IResponse;
-import brawljars.common.RawResponse;
-
 import com.google.gson.annotations.SerializedName;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+
+import supercell.api.wrapper.essentials.common.IResponse;
+import supercell.api.wrapper.essentials.common.RawResponse;
 
 @Data
 @EqualsAndHashCode(callSuper = true)