+++ src/main/java/brawljars/api/intern/players/PlayerApiImpl.java
@@ -16,12 +16,13 @@
  */
 package brawljars.api.intern.players;
 
-import brawljars.api.ApiContext;
-import brawljars.api.BaseApi;
 import brawljars.api.intern.players.battlelog.BattleLogRequest;
 import brawljars.api.intern.players.battlelog.BattleLogResponse;
 import brawljars.api.intern.players.info.PlayerRequest;
 import brawljars.api.intern.players.info.PlayerResponse;
+
+import supercell.api.wrapper.essentials.api.ApiContext;
+import supercell.api.wrapper.essentials.api.BaseApi;
 
 import java.util.concurrent.Future;