+++ src/main/java/brawljars/api/intern/rankings/RankingApiImpl.java
@@ -16,8 +16,6 @@
  */
 package brawljars.api.intern.rankings;
 
-import brawljars.api.ApiContext;
-import brawljars.api.BaseApi;
 import brawljars.api.intern.rankings.brawler.BrawlerRankingsRequest;
 import brawljars.api.intern.rankings.brawler.BrawlerRankingsResponse;
 import brawljars.api.intern.rankings.club.ClubRankingsRequest;
@@ -28,6 +26,9 @@ import brawljars.api.intern.rankings.powerplay.PowerplayRankingsRequest;
 import brawljars.api.intern.rankings.powerplay.PowerplayRankingsResponse;
 import brawljars.api.intern.rankings.powerplay.PowerplayRankingsSeasonsRequest;
 import brawljars.api.intern.rankings.powerplay.PowerplayRankingsSeasonsResponse;
+
+import supercell.api.wrapper.essentials.api.ApiContext;
+import supercell.api.wrapper.essentials.api.BaseApi;
 
 import java.util.concurrent.Future;