+++ src/main/java/brawljars/api/intern/rankings/club/ClubRankingsRequest.java
@@ -16,9 +16,9 @@
  */
 package brawljars.api.intern.rankings.club;
 
-import brawljars.common.PaginationRequest;
-
 import lombok.Builder;
+
+import supercell.api.wrapper.essentials.common.PaginationRequest;
 
 import java.util.Map;